Nullsafety update

+++ lib/src/cbor_constants.dart
@@ -36,10 +36,10 @@ const int ai27 = 27;
 const int aiBreak = 31;
 
 /// Numerical constants
-final int two8 = pow(2, 8);
-final int two16 = pow(2, 16);
-final int two32 = pow(2, 32);
-final int two64 = pow(2, 64);
+final int two8 = pow(2, 8).toInt();
+final int two16 = pow(2, 16).toInt();
+final int two32 = pow(2, 32).toInt();
+final int two64 = pow(2, 64).toInt();
 const double halfLimitUpper = 65504.0;
 const double halfLimitLower = -65504.0;
 const double singleLimitUpper = 3.4028234663852886e38;
